+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+## 0.28.1-rc.0:
+- Add TOML Jinja2 Filters (#534)
+- Fix issues #535 and #536 by resolving relative paths in the copy input type and not clobbering the output dir (#539)
+
+## 0.28.0:
+- Update reclass and some pip dependencies. (#528)
+- Allow using .yaml output file extension along with .yml (#525)
+  - Note: if you have .yaml files in your inventory instead of .yml, now kapitan should recognise them as expected. If you see changes in your compiled output just rename your .yaml files to .yml to keep the compiled diff identical.
+- Fix embedded refs encoding value (#523)
+- Add new Helm dependency input type (#520)
+- adding dir_files_list and dir_files_read in jsonnet (#519)
+- Add file_exists function to jsonnet (#518)
+
+## 0.28.0-rc.2:
+- Update reclass and some pip dependencies. (#528)
+
+## 0.28.0-rc.1:
+- Allow using .yaml output file extension along with .yml (#525)
+
+## 0.28.0-rc.0:
+- Fix embedded refs encoding value (#523)
+- Add new Helm dependency input type (#520)
+- adding dir_files_list and dir_files_read in jsonnet (#519)
+- Add file_exists function to jsonnet (#518)
+
 ## 0.27.4:
 - Fix linter argument flag to refs-path instead of deprecated secrets-path (#516)